+++ config-model/src/main/java/com/yahoo/searchdefinition/FeatureNames.java
@@ -5,13 +5,10 @@
  */
 package com.yahoo.searchdefinition;
 
-import com.yahoo.searchlib.rankingexpression.rule.ReferenceNode;
+import com.yahoo.searchlib.rankingexpression.Reference;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.Optional;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 /**
  * Utility methods for query, document and constant rank feature names
@@ -22,16 +19,16 @@ public class FeatureNames {
 
     private static final Pattern identifierRegexp = Pattern.compile("[A-Za-z0-9_][A-Za-z0-9_-]*");
 
-    public static ReferenceNode.Reference asConstantFeature(String constantName) {
-        return ReferenceNode.Reference.simple("constant", quoteIfNecessary(constantName));
+    public static Reference asConstantFeature(String constantName) {
+        return Reference.simple("constant", quoteIfNecessary(constantName));
     }
 
-    public static ReferenceNode.Reference asAttributeFeature(String attributeName) {
-        return ReferenceNode.Reference.simple("attribute", quoteIfNecessary(attributeName));
+    public static Reference asAttributeFeature(String attributeName) {
+        return Reference.simple("attribute", quoteIfNecessary(attributeName));
     }
 
-    public static ReferenceNode.Reference asQueryFeature(String propertyName) {
-        return ReferenceNode.Reference.simple("query", quoteIfNecessary(propertyName));
+    public static Reference asQueryFeature(String propertyName) {
+        return Reference.simple("query", quoteIfNecessary(propertyName));
     }
 
     /**
@@ -39,7 +36,7 @@ public class FeatureNames {
      * or empty if it is not a valid query, attribute or constant feature name
      */
     public static Optional<String> argumentOf(String feature) {
-        Optional<ReferenceNode.Reference> reference = ReferenceNode.Reference.simple(feature);
+        Optional<Reference> reference = Reference.simple(feature);
         if ( ! reference.isPresent()) return Optional.empty();
         if ( ! ( reference.get().name().equals("attribute") ||
                  reference.get().name().equals("constant") ||
